Update per #84 review; tests, and dist code
