+++ pymc_experimental/tests/distributions/__init__.py
@@ -1,0 +1,2 @@
+from pymc_experimental.distributions import histogram_utils
+from pymc_experimental.distributions.histogram_utils import histogram_approximation
