+++ pymc_experimental/tests/distributions/test_continuous.py
@@ -1,0 +1,108 @@
+
+#   Copyright 2020 The PyMC Developers
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
+# general imports
+import numpy as np
+import pytest
+import scipy.stats as st
+import scipy.stats.distributions as sp
+
+import pymc as pm
+
+# test support imports from pymc
+from pymc.tests.distributions.util import (
+    BaseTestDistributionRandom,
+    R,
+    Rplus,
+    Domain,
+    check_logp,
+    check_logcdf,
+    assert_moment_is_expected,
+    seeded_scipy_distribution_builder,
+)
+
+# the distributions to be tested
+from pymc_experimental.distributions import (
+    GenExtreme,
+)
+
+class TestGenExtreme():
+    """
+    Wrapper class so that tests of experimental additions can be dropped into
+    PyMC directly on adoption.
+    """
+
+    def test_genextreme():
+        check_logp(
+            GenExtreme,
+            R,
+            {"mu": R, "sigma": Rplus, "xi": Domain([-1, -1, -0.5, 0, 0.5, 1, 1])},
+            lambda value, mu, sigma, xi: sp.genextreme.logpdf(value, c=-xi, loc=mu, scale=sigma),
+        )
+        check_logcdf(
+            GenExtreme,
+            R,
+            {"mu": R, "sigma": Rplus, "xi": Domain([-1, -1, -0.5, 0, 0.5, 1, 1])},
+            lambda value, mu, sigma, xi: sp.genextreme.logcdf(value, c=-xi, loc=mu, scale=sigma),
+        )
+
+    @pytest.mark.parametrize(
+        "mu, sigma, xi, size, expected",
+        [
+            (0, 1, 0, None, 0),
+            (1, np.arange(1, 4), 0.1, None, np.arange(1, 4) * (1.1 ** -0.1 - 1) / 0.1),
+            (np.arange(5), 1, 0.1, None, np.arange(5) + (1.1 ** -0.1 - 1) / 0.1),
+            (
+                0,
+                1,
+                np.linspace(-0.2, 0.2, 6),
+                None,
+                ((1 + np.linspace(-0.2, 0.2, 6)) ** -np.linspace(-0.2, 0.2, 6) - 1)
+                / np.linspace(-0.2, 0.2, 6),
+            ),
+            (1, 2, 0.1, 5, np.full(5, 1 + 2 * (1.1 ** -0.1 - 1) / 0.1)),
+            (
+                np.arange(6),
+                np.arange(1, 7),
+                np.linspace(-0.2, 0.2, 6),
+                (3, 6),
+                np.full(
+                    (3, 6),
+                    np.arange(6)
+                    + np.arange(1, 7)
+                    * ((1 + np.linspace(-0.2, 0.2, 6)) ** -np.linspace(-0.2, 0.2, 6) - 1)
+                    / np.linspace(-0.2, 0.2, 6),
+                ),
+            ),
+        ],
+    )
+    def test_genextreme_moment(mu, sigma, xi, size, expected):
+        with pm.Model() as model:
+            GenExtreme("x", mu=mu, sigma=sigma, xi=xi, size=size)
+        assert_moment_is_expected(model, expected)
+
+
+class TestGenExtreme(BaseTestDistributionRandom):
+    pymc_dist = GenExtreme
+    pymc_dist_params = {"mu": 0, "sigma": 1, "xi": -0.1}
+    expected_rv_op_params = {"mu": 0, "sigma": 1, "xi": -0.1}
+    # Notice, using different parametrization of xi sign to scipy
+    reference_dist_params = {"loc": 0, "scale": 1, "c": 0.1}
+    reference_dist = seeded_scipy_distribution_builder("genextreme")
+    tests_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_pymc_draws_match_reference",
+        "check_rv_size",
+    ]
