--- pymc_experimental/tests/test_distributions_random.py
@@ -1,38 +1,0 @@
-#   Copyright 2020 The PyMC Developers
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
-# general imports
-
-# test support imports from pymc
-from pymc.tests.test_distributions_random import (
-    BaseTestDistribution,
-    seeded_scipy_distribution_builder,
-)
-
-# the distributions to be tested
-import pymc_experimental as pmx
-
-
-class TestGenExtreme(BaseTestDistribution):
-    pymc_dist = pmx.GenExtreme
-    pymc_dist_params = {"mu": 0, "sigma": 1, "xi": -0.1}
-    expected_rv_op_params = {"mu": 0, "sigma": 1, "xi": -0.1}
-    # Notice, using different parametrization of xi sign to scipy
-    reference_dist_params = {"loc": 0, "scale": 1, "c": 0.1}
-    reference_dist = seeded_scipy_distribution_builder("genextreme")
-    tests_to_run = [
-        "check_pymc_params_match_rv_op",
-        "check_pymc_draws_match_reference",
-        "check_rv_size",
-    ]
